+++ dialbb/builtin_blocks/stn_management/stn_manager.py
@@ -80,6 +80,8 @@ CONTEXT_KEY_SUB_DIALOGUE_STACK: str = '_sub_dialogue_stack'
 CONTEXT_KEY_REACTION: str = '_reaction'
 CONTEXT_KEY_REQUESTING_CONFIRMATION: str = '_requesting_confirmation'
 CONTEXT_KEY_TURNS_IN_STATE: str = '_turns_in_state'
+CONTEXT_KEY_SESSION_ID: str = '_session_id'
+CONTEXT_KEY_USER_ID: str = '_user_id'
 
 INPUT_KEY_AUX_DATA: str = "aux_data"
 INPUT_KEY_SENTENCE: str = "sentence"
@@ -358,6 +360,8 @@ class Manager(AbstractBlock):
                            CONTEXT_KEY_SUB_DIALOGUE_STACK: [],
                            CONTEXT_KEY_REACTION: "",
                            CONTEXT_KEY_TURNS_IN_STATE: 1,
+                           CONTEXT_KEY_SESSION_ID: session_id,
+                           CONTEXT_KEY_USER_ID: user_id,
                            CONTEXT_KEY_REQUESTING_CONFIRMATION: False}
 
                 self._add_context(session_id, context)