+++ Changelog.md
@@ -1,9 +1,19 @@
 # Changelog
 
 
-## 1.1.2 (2025.11.23)
+## next version
 
 - Conforms to Python 3.12
+
+- ChatGPT Dialogue Builtin Block
+
+  - Allows to use gpt-5* other than gpt-5
+
+- STN Manager Embedded Block
+
+  - Allows to use gpt-5* other than gpt-5
+
+  - Allows to log in scenario functions
 
 ## 1.1.1 (2025.11.22)